Try to fix jest environment issues & addDelete.Test.js file
