+++ package.json
@@ -4,7 +4,7 @@
   "description": "- [📖 About the Project](#about-project)\r   - [🛠 Built With](#built-with)\r     - [Tech Stack](#tech-stack)\r     - [Key Features](#key-features)\r - [💻 Getting Started](#getting-started)",
   "private": true,
   "scripts": {
-    "test": "jest --coverage",
+    "test": "jest",
     "watch": "webpack --watch",
     "start": "webpack serve --open",
     "build": "webpack"
@@ -13,14 +13,17 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/core": "^7.22.11",
+    "@babel/preset-env": "^7.22.10",
     "babel-eslint": "^10.1.0",
+    "babel-jest": "^29.6.4",
     "css-loader": "^6.8.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.28.0",
     "hint": "^7.1.10",
     "html-webpack-plugin": "^5.5.3",
-    "jest": "^29.6.4",
+    "jest-environment-jsdom": "^27.5.1",
     "style-loader": "^3.3.3",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
@@ -375,6 +378,7 @@
     "isobject": "^3.0.1",
     "jackspeak": "^2.3.0",
     "jake": "^10.8.7",
+    "jest": "^27.5.1",
     "jest-worker": "^27.5.1",
     "js-library-detector": "^6.7.0",
     "js-tokens": "^4.0.0",
