+++ src/modules/GrabUserInput.js
@@ -5,14 +5,24 @@ const todoinput = document.querySelector('.todoinput');
 
 class GrabUserInput {
   grab = (val) => {
+    // val.preventDefault();
+    // if (todoinput.value) {
+    //   const idx = JSON.parse(localStorage.getItem('todos')).length + 1;
+    //   const todo = new Todo(todoinput.value, idx);
+    //   const todosArr = JSON.parse(localStorage.getItem('todos'));
+    //   TodoItems.addtodo(todo, todosArr);
+    //   return;
+    // }
+    // todoinput.value = '';
     val.preventDefault();
-    if (todoinput.value) {
-      const idx = JSON.parse(localStorage.getItem('todos')).length + 1;
-      const todo = new Todo(todoinput.value, idx);
-      const todosArr = JSON.parse(localStorage.getItem('todos'));
-      TodoItems.addtodo(todo, todosArr);
+    if (todoinput.value.trim() === '') {
       return;
     }
+
+    const idx = JSON.parse(localStorage.getItem('todos')).length + 1;
+    const todo = new Todo(todoinput.value.trim(), idx);
+    const todosArr = JSON.parse(localStorage.getItem('todos'));
+    TodoItems.addtodo(todo, todosArr);
     todoinput.value = '';
   }
