+++ src/modules/__tests__/addDelete.test.js
@@ -1,0 +1,45 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import todomarkup from '../TodoMarkup.js';
+import TodoItems from '../TodoItems.js';
+
+let store = [{
+  description: 'laundry',
+  completed: false,
+  index: 1,
+}];
+
+document.body.innerHTML = '<div class=\'list-items\'></div>';
+
+// add
+describe('add a new todoitem', () => {
+  it('adds a new object to the store', () => {
+    const number = store.length;
+    TodoItems.addtodo({
+      description: 'read a book',
+      completed: false,
+      index: 2,
+    }, store);
+    expect(store.length).toBe(number + 1);
+  });
+
+  it('adds a node to the DOM', () => {
+    const todoList = document.querySelector('.list-items');
+    let markup = '';
+    store.forEach((todo) => {
+      markup += todomarkup(todo);
+    });
+    todoList.innerHTML = markup;
+    expect(markup).toEqual(document.body.children[0].innerHTML);
+  });
+
+  it('checks for a truthy value', () => {
+    expect(TodoItems.addtodo({
+      description: 'wash the dishes',
+      completed: false,
+      index: 3,
+    }, store)).toBeTruthy();
+  });
+});
